requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pypdf
python-multipart
openai
